Allow running tests/binary in debug mode
Running tests or the binary in debug mode would previously panic when
trying to subtract 1 from `0usize` in the grid problem of Day 12. This
was known/expected. This is annoying though. It's nice to run in debug
mode so you can _feel_ the slow points. `wrapping_sub` is free in
release mode since it's the default for subtraction. This should allow
me to feel more pain and write better code.